requirements.txt
discord.py==2.1.1
table2ascii==1.1.0
topggpy~=1.4.0
configparser~=5.3.0
asyncio